+++ src/test/editor-service/editor-service_test.ts
@@ -16,12 +16,12 @@ import {assert} from 'chai';
 import * as fs from 'fs';
 import * as path from 'path';
 
-import {AttributesCompletion, EditorService, ElementCompletion, Severity, Warning} from '../../editor-service/editor-service';
+import {AttributesCompletion, EditorService, ElementCompletion} from '../../editor-service/editor-service';
 import {LocalEditorService} from '../../editor-service/local-editor-service';
 import {RemoteEditorService} from '../../editor-service/remote-editor-service';
 import {FSUrlLoader} from '../../url-loader/fs-url-loader';
 import {PackageUrlResolver} from '../../url-loader/package-url-resolver';
-
+import {Severity, Warning} from '../../warning/warning';
 import {invertPromise} from '../test-utils';
 
 function editorTests(editorFactory: (basedir: string) => EditorService) {
